Introduced a new operator, OpSubstringQuantify.  This yields a sequence of tuples containing adjacent elements in its sequence operand.  A rewrite rule has been added that rewrites expressions that quantify over sequences using indexing where the indices are used to select adjacent elements.  For example: `forAll i int(2..5) . constraint(s(i-1),s(i))`

+++ src/operators/opPowerSet.h
@@ -10,7 +10,7 @@ struct OperatorTrates<OpPowerSet> {
     struct OperandTrigger;
 };
 struct OpPowerSet : public SimpleUnaryOperator<SetView, SetView, OpPowerSet> {
-    int sizeLimit;
+    size_t sizeLimit;
     OpPowerSet(ExprRef<SetView> operand, int sizeLimit = -1)
         : SimpleUnaryOperator<SetView, SetView, OpPowerSet>(std::move(operand)),
           sizeLimit(sizeLimit) {
